MDEV-23326: mysql_tzinfo_to_sql add --skip-wsrep-check
The --skip-write-binary-log added to mysql_tzinfo_to_sql in
MDEV-18778 was only effective if galera was enabled on the server.
This is because it tied together three concepts under on option,
binary logging wsrep replication and using innodb as a transitional
table type.

There are two scenarios where the behaviour of --skip-write-binary-log
would be very useful without being dependant on galera.

1) For loading timezones in to a classical async replication master
without the timezones being written to a binary log. Here we'd
like the --skip-write-binary-log, binary logs are explictly
skipped, but without the wsrep checks..

2) Loading timezone tables with innodb is signifcantly faster
than Aria (with TRANSACTIONAL=1). Slow storage particularly
in container environments (the main MDEV-23326 problem) can
cause excessive initialization time. We'd like to improve
the speed here regardless of if its a Galera environment or not.

To address case 2), the lack of checking the wsrep status also
make it possible to use mysql_tzinfo_to_sql in the mysqld --bootstrap
input allowing for a single initialization of all system tables.

In cleaning up part of MDEV-18778, when changing back from InnoDB
to MyISAM, include the ORDER BY, and don't include this if only
loading leap seconds.