WIP: Remove mlog_checkpoint_lsn and more (breaks recovery further)
This is a temporary development step, to bring the recovery code closer
to the log records that are being written out.

Ultimately, we will introduce a new redo log format tag for the
two-file format, and we might retain code for crash recovery from
the 10.5 format. If we do that, then this code should be mostly retained.