MDEV-24302 RESET MASTER hangs in innodb.group_commit_crash_no_optimize_thread
The failure was caused by a race between
binlog checkpoint requester thread and a notifier one for
the value of request list.
*After* creating and storing the request to a list, the requester did not
check, whether a notification condition has been taken place.
And it had to because a potential notifier may have read NULL value of
the list in `innobase_mysql_log_notify`.
The request therefore would be dangling in the list to hold on attempts
to remove respective binlog files.
The failure is likely to be 10.5 specific where log flushing was optimized
it's fixed in early versions too.

The patch makes sure the race can't occur anymore through
A. the requester checks flush_lsn once again upon the list pointer
update and when the value suggests - `lsn == flush_lsn` -
it triggers notification at once by itself;
B. the list pointer is made std::atomic, the relaxed
memory order suffices to correctly update/read-from it.