MDEV-24526 binlog rotate via FLUSH LOGS may obsolate binlog file too eary
There was a race between a committing transaction and a following in binlog
order FLUSH LOGS that could race over to log a 2nd Binlog checkpoint event
in the new file *before* the logged-in-old-binlog transaction gets committed in
Innodb.
Such behaviour may ensue server crash-recovery error in losing
the transaction should the server crash before it eventually gets
committed in the Engine.

The race is tackeled by enforcing the necessary set of mutexes to be acquired
by FLUSH-LOGS handler in the correct order (of the group commit leader
pattern).

Note, there remain two cases where a similar race is still possible:
  - the above race as it is when the server is run with ("unlikely")
    non-default `--binlog-optimize-thread-scheduling=0` (MDEV-24530), and
  - at unlikely event of bin-logging of Incident event (MDEV-24531) that
    also triggers binlog rotation,
    in both cases though with lesser chances after the current fixes.