Aria: Add transaction id to log of create table
In the case of a crash directly after a creation of an Aria table,
Aria recovery would think that the table was from another system and
require a repair of the table and inform that the table is 'zerofilled".
This would cause no harm, but was confusing to see when testing atomic
alter table.

Fixed by logging the create transaction id to the log.

Other things:
- aria_chk --describe now prints a warning if the table was moved from
  another system.