Aria: Add transaction id to log of create table
In the case of a crash directly after a creation of an Aria table,
Aria recovery would think that the table was from another system and
require a repair of the table and inform that the table is 'zerofilled".
This would cause no harm, but was confusing to see when testing atomic
alter table.

Fixed by logging the create transaction id to the log.

Other things:
- Added "show table status from sys" to maria_empy_logs. This ensures one
  does not get any zerofill warnings when sys/sys_config is used by other
  tests.
- aria_chk --describe now prints a warning if the table was moved from
  another system.
- Logging of truncate (maria_delete_all_rows) is changed to use the
  current trid for the create table.
  This is to ensure that we do not run into the same problem with truncate.
- Changed back sys_config table to Aria as this patch should fix the
  "zerofill" problem in buildbot.
- Added scripts/mysql_sys_schema.sql to .gitignore