MDEV-25395 server recovery hits replication event checksum error
When @@global.binlog_checksum is changing from none to
crc32 and only the first of two Binlog_checkpoint_log_event gets written to
the crc32 rotated binlog file before the server crashes, the recovery faces a checksum
verification error.

The reason was that recovery algorithm did not realize
potentially varying binlog checksum at  scanning of
a sequence of binlog files.
That is corrected now. A simulation test confirms the fixes.

rpl_cant_read_event_incident got affected because the test
adopts an external not properly closed binlog file that contains
events which can't pass checksum verification.
A work-around is done to accept that binlog and evade recovery.